backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
scikit-learn
tensorflow
sqlalchemy
python-multipart
requests
httpx
jinja2
python-dotenv
google-genai
